feat: after merge

+++ src/main/java/com/example/part35teammonew/domain/userActivity/Dto/ArticleInfoView.java
@@ -24,7 +24,7 @@ public class ArticleInfoView {
   private String articleSummary;
 
   private int articleCommentCount;
-  private int articleViewCount;
+  private Long articleViewCount;
 
   @Override
   public boolean equals(Object o) {
